debug views

+++ stats/stats-server/src/main/java/ru/practicum/statistics_service/model/ViewStats.java
@@ -10,5 +10,5 @@ import lombok.Data;
 public class ViewStats {
     private String app;
     private String uri;
-    private long hits;
+    private Integer hits;
 }
